Fix error handing in Tornado parser.
The second argument to tornado.web.HTTPError, log_message, is expected
to be a string, but TornadoParser#handle_error was passing an Exception
object instead, which caused Tornado applications to time out. This
patch sends the message of the exception rather than the exception
itself, and adds a test case that also times out if the wrong argument
is passed to HTTPError.