Implement get_args_by_location for all parsers
Add implementations of this method for all existing parsers. For the
most part, the pattern is quite simple and nearly identical. Usually,
JSON data-loading needs to be moved into a helper for use both in
get_arbs_by_location and parse_json.

The asyncparser is particularly significant, as it reimplements much of
the core parser in order to make get_args_by_location an async
(awaitable) method.