Improvements per @lafrech's review
1. Fix a typo in the quickstart doc

Rewrite "content_type" to "user_type". Just a slip-up.

2. Flatten out the nesting in MultiDictProxy.__getitem__

Although this was flagged as an `else-return` which could flatten
out to just a return, Jérôme also suggested a version which treats the
outer condition as an `if ...: return` so that we can dedent the rest
of the logic.
The end-result is easier to read, so let's go with that.

3. Add a test for passing a single value to a List field

This applies to all parsers, but it was required by FalconParser in
particular. If you remove MultiDictProxy from the presentation of
query params, headers, etc. from FalconParser, it works in most cases.
However, because field values only get wrapped in lists by that parser
if there are multiple values supplied, we need some intermediary to
understand that lookups for List fields always need a `list`-type
value. MultiDictProxy serves that purpose, as a schema-aware dict type.