Change DelimitedList to always go to/from a string
Parsing/deserializing a DelimitedList will now insist that the input
is a string or bytes. List-like values for DelimitedList are no longer
valid.
Also, for symmetry, remove the `as_string` parameter -- always
serialize DelimitedList as a string.

As a final change, make the default error message for an invalid
delimited list slightly different from the one for List, to
disambiguate things.

closes #423