Add a warning if validation errors are suppressed
We can't be certain that the user doesn't know what they're doing in
this case. It's conceivable that they have a validation hook which
doesn't raise an error and they're prepared for `parse()` to return
`None`. However, in that case, it would be beneficial to at least emit
a warning, because this is likely to confuse users.

resolves #518