iter(MultiDictProxy) checks if contents are tuples
Specifically in order to handle headers more gracefully,
MultiDictProxy needs to be careful about how it exposes __iter__.
Marshmallow gets a list of keys in the data it's loading by calling
`set(data)`, which implicitly calls `iter()`. And MultiDictProxy was
naively proxying that method to the underlying object's `__iter__`.

However, for Flask Headers and potentially other specialized request
objects, `iter()` does not produce an iterator over keys, but over
some other type. In the case of Flask Headers, the type produced is a
tuple of (key, value).

Rather than trying to make MultiDictProxy aware of all of the
dict-like types it might be handling, simply wrap the `iter()` call in
a loop which checks if objects are tuples. If they are, pop the first
element. In the case of headers, at least, that is the key.

This can definitely be made to fail on oddly-constructed underlying
data. e.g. If you define a type where iter yields an empty tuple,
this will blow up. However, no known real data from frameworks is
shaped that way.

fixes #500