ValueError if error handler does not reraise
In 6.x this was only a warning. Now, if someone defines an error
handler which does not raise an exception, webargs will raise a
ValueError. This prevents use_args/use_kwargs/parse from returning
malformed or unexpected data if an error handler fires but does not
raise an exception.

Several tests needed tweaking to handle this, as they defined handlers
which did not raise errors.

closes #527