Remove support for marshmallow2
Anywhere that version dispatch was used, switch to the ma3 version of
the code. Tests skipped on marshmallow3 are dropped.

Remove dict2schema, since it is now redundant with Schema.from_dict .

Update the changelog and docstrings around `unknown=...` to remove
note that it only works on marshmallow3. Remove any documentation
which notes a marshmallow2 (vs ma3) discrepancy or a way of getting
the same behavior out of both versions.

Update readme badge to ma3 only.

tox and azure pipelines config are altered to drop marshmallow2
builds.

Update `setup.py` to require `marshmallow>=3.0.0`

closes #540