Update min django version, comment on load_headers
We now support Django 2.2.0 or later (for `HttpRequest.headers` to be
available). Even though the requirement won't typically be passed on
to users, it's the only place where webargs lists its (expected)
compatible version ranges.

Also add a comment to DjangoParser.load_headers which notes explicitly
that the headers object isn't a multidict, and therefore isn't
proxied.