Test against minimum supported framework versions
This adds a tox factor, `mindeps`, which specifies explicitly that we
want to test with old versions of the various frameworks.
Right now, it is only specified for our minimum and maximum supported
python versions -- meaning 3.6 and 3.9 -- just to keep the build
matrix a little bit smaller.

This testing is meant to "catch us" if we add a chunk of code to one
of the parsers which will break on older framework versions. For
example, we might use a request attribute without realizing that it was
added in a minor version of a framework. We'd then  break
compatibility with older (possibly still supported) framework
versions.

This explicitly does not add any public documentation about minimum
compatible versions. We don't offer a new guarantee to users and the
support we offer is still loose and "best effort". But it does update
the versions listed in `setup.py`, so if people are looking there to
get a feel for what we support, they'll see something reasonable.

flask and aiohttp both failed on these builds. flask because it didn't
start pinning the version of werkzeug until 0.12.5, and aiohttp
because versions prior to 3.0.8 aren't py3.9-compatible. So these have
been updated.