Fix test order sensitivity
It's not possible to construct an asyncio Future (which tornado Future
aliases) without an event loop present.
I believe that what was going wrong with tests being order-sensitive
before was that if aiohttp tests ran before tornado tests, the loop
fixture teardown would leave us in a state in which the tornado tests
could not run.

In any event, aiohttp's pytest plugin provides an event loop as a
context-manager fixture, so using it ensures that there is a valid
event loop for tornado to use.

With this change in place, the test_py3 modules can be moved up to the
same level as the rest of the tests.