Add type annotations to Parser methods
A lot of these annotations are copy-paste from the types for
AsyncParser, but there are a few odd cases which require more
explanation or special handling.

One thing which existed before were a number of `Union[None, foo]`
types which I have switched to `Optional[foo]` to improve readability.
We always annotate explicitly when things are Optional so that users
doing strict checking in `mypy` won't get spurious errors.

`_callable_or_raise` is annotated to preserve the type of its input.
However, this ran afoul of some incorrect handling of `callable()` in
mypy. I've filed https://github.com/python/mypy/issues/8974 and put in
place a workaround with a comment.

But things get very hairy in the AsyncParser subclass, since it now
has different types from its parent class (and mypy type checking can
see that it does). In particular, an error handler for the async
parser may be async. So its return type isn't `NoReturn` but an
`Awaitable[NoReturn]`. To handle this, check if an error handler is
async, cast it as an awaitable or not depending on whether or not it
is a coroutine (`inspect` check) and await it if necessary.
This blows up with an error about an implicit return on a function
annotated as NoReturn. Which... it turns out this is a mypy issue:
https://github.com/python/mypy/issues/8974 . To workaround, an
unreachable `raise NotImplementedError("unreachable")` "forces" mypy
to understand how the function behaves.

Another bit of weirdness is that the signature for
`_on_validation_error` in AsyncParser breaks type checking because it
doesn't match `Parser._on_validation_error`. Since this is a purely
internal hook, rename as `_async_on_validation_error` to patch over
this hole.

Also mark `schema: ma.Schema` in multidictproxy, but don't worry about
annotating it beyond that.