Fix annotations which weren't caught by pre-commit
When mypy runs individually on various files, but without webargs and
marshmallow installed, it can't do full type checking because the
types from marshmallow are not available to it. It also seems to
result in some failures to catch errors split between modules within
webargs.

This change ensures that mypy will pass when run on webargs in this
kind of environment. First, it fixes #569 by ensuring that keyword
args to `_serialize` in `DelimitedFieldMixin` are passed through
correctly. Second, several annotations are edited for accuracy, and
`aiohttpparser` is no longer annotated with `aiohttp.web.Request` as
the argument type for requests. This caused errors because it
conflicts with the `Request` generic used in the base Parser.