Improve mypy tox target and precommit conf
The issues with undiscovered type errors actually derived from the
fact that `marshmallow` was not available in the `mypy` environment.
As a result, values like `missing` could not be resolved and were
typed as `Any`. That, in turn, leads to "strangely incorrect"
signatures like `Union[Dict, Any]` (a.k.a. `Any`), which then can
result in type errors that ought to be identified being missed.

Once `marshmallow` is installed, mypy will find types like
`Union[Dict, _Missing]`, and can determine that they do (or do not)
match as appropriate.

However, the tox target and CI build can be used for a more extensive
test of webargs with all frameworks installed. As a result, keeping
both is valuable, and this change improves both configs to catch a
wider class of errors.