Add support for the `is_multiple` field attribute
This resolves #563

Detection of fields as `is_multiple` for multi-dict unpacking now
allows custom fields to use an attribute `is_multiple=T/F` to control
the behavior directly. This allows for a subclass of `List` to set
`is_multiple=False` (done by `DelimitedList`) and allows non-`List`
subclasses to be detected as `multiple`.

Because `is_multiple` is an attribute, it can be set on the field
class or on an instance.

A noted TODO comment questions whether or not `ma.fields.Tuple` should
be added to the auto-detected fields for `is_multiple`.

Tests cover
- a field with is_multiple=True
- a field with is_multiple=False
- a field where is_multiple is not set
- a List(Str()) field

and run against the collection of multidict implementations used to
test MultiDictProxy.