Make KNOWN_MULTI_FIELDS configurable and add doc
Rather than a module-level constant, move ``KNOWN_MULTI_FIELDS`` to be
a parser attribute. This will allow users to override or modify it.
However, moving it creates a problem: whereas previously `is_multiple`
was just a function which MultiDictProxy could safely call, now it's
a method and requires a Parser instance.

To handle this, change the relationship between the multidictproxy and
core modules. `core` now requires `multidictproxy` and the base
`Parser` provides a private method for building a MultiDictProxy
object.

MultiDictProxy doesn't depend on parsers (tried that, didn't like it).
Instead, a MultiDictProxy can be given a `known_multi_fields` argument
when created. The `_makeproxy` helper on the Parser class just makes
sure to pass `known_multi_fields=self.KNOWN_MULTI_FIELDS`, and
everything falls into place from there. Numerous tweaks to parser
classes use `_makeproxy` instead of directly instantiating a
MultiDictProxy, but that's the bulk of the changeset.

New docs in a section towards the end of the advanced doc explains how
multi field detection works and explains how you can set
`is_multiple = True` or add a field class to
`Parser.KNOWN_MULTI_FIELDS`.