Narrow type of argmap from Mapping to Dict
The `Mapping` case is meant to support passthrough to
`Schema.from_dict`. However, only `Dict` is actually supported.
Therefore, narrowing this from `Mapping` to `Dict` makes the
annotation more accurate.

In terms of runtime behavior, the check is changed from looking at
`isinstance(argmap, collections.abc.Mapping)` to
`isinstance(argmap, dict)`. This should not impact any users and
harmonizes the logic with the declared types and documentation.

Technically, it is possible to make a non-dict mapping type which is
close enough to dict to be acceptable for `Schema.from_dict`. Such a
usage is unlikely and was not supported intentionally in the first
place.

To avoid unclear and surprising behaviors, non-dict mapping types will
now trigger a `TypeError`. The test case for this uses a UserDict,
which is dict-like but not a subclass of dict, to demonstrate the sort
of case this will catch.

Additionally, fix some minor issues with some of the type annotations.
In order for a decorator to preserve the type of the decorated
function, it needs to be annotated with a type variable. There were a
couple of cases of this in `Parser` which have been fixed.