Pin markupsafe and itsdangerous in mindeps build
Because `flask` does not pin its dependencies, it is necessary to
include at least some pins when trying to install any version other
than the latest. A comment in the tox config explains this briefly.

In order to also creep forward the lowest supported version (0.12.5 is
*very* old), use one of the earlier flask 1.x releases, along with
versions of markupsafe and itsdangerous which are approximately
contemporaneous with it.

Whether or not these versions are the best choice, this set works and
should get the build passing again.