Combine AsyncParser with core.Parser
Combine the implementations. Because AsyncParser already checks if
it's being called on coroutines or synchronous functions, this is
mostly a matter of "upgrading" the parser implementation to the
AsyncParser one. The only method which *must* be split is `parse`
(which cannot be both sync and async), so handle this by having two
method names: parse and async_parse.

By a similar token, we need sync and async variants of a couple of
helpers. However, try to consolidate and reduce duplication between
components. `parse` and `async_parse` use some shared helpers for
this, and there's some minor shuffling to do this with
_load_location_data and _async_load_location_data so that they don't
overlap more than necessary.

Tests need a bit of adjustment for this. Some new async tests in the
core tests are added, and the flask async tests should now pass. In
order to allow tests to run under "mindeps", the testsuite parses the
flask version number and skips async flask tests when flask<2

Some notable results of these changes:

- AsyncParser turns into a tiny shim class!
  It just defines `parse` as an alias for `async_parse`. It could be
  removed in a future major version.

- FlaskParser (and any other framework parser with async support) will
  now work on async views with no modifications. The needed dispatch
  is inherited from the base parser.

- It will still be pretty easy to override `parse` in a subclass, but
  overriding `use_args` while preserving the sync/async behavior gets
  trickier.

- The amount of duplication between async/non-async codepaths is
  reduced! It isn't 0, and probably will never be completely free of
  duplication/repetition, but it's improved.