Add `empty` as a parameter to DelimitedFieldMixin
`empty` allows more graceful support for non-string delimited data,
like lists of ints.

`empty` is annotated on the DelimitedFieldMixin class but is only set
on its subclasses. This technically means that a custom subclass of
the mixin class could be defined by a third party and would be broken
by this change, but this is acceptable.

resolves #823