Make Parser a generic type over Request
`webargs.core.Request` was declared as a TypeVar but then never
bound to a relevant class-scope. As a result, it effectively evaluated
as `Any` in all contexts.
Not only does this make typing less useful for us within webargs, it
also means that users have a less clear typing contract to satisfy.

To resolve this, `Request` needs to be bound to the Parser class
scope by means of the class becoming `Generic[Request]`. Subsequent to
this, it becomes possible to add the relevant Request class for
each framework parser as part of the type definition.

Throughout parser methods, like location loaders, every usage of the
request object can be annotated with `webargs.core.Request`. This
means the same thing as the relevant concrete request type for the
framework parsers.

Finally, having made this change, it is possible to circle back to
issue #836 and make the `ArgMap` type into a `Mapping` and have
its new type propagate correctly. Note that `ArgMap` is a type alias
which includes the `Request` type, meaning that it is an alias which
is only valid in a context where `Request` is bound -- i.e. inside of
a Parser class.

As a minor behavioral tweak required by this change, when `ArgMap` is
passed to `Schema.from_dict`, it must be explicitly dict-ified if it
is a non-dict Mapping. `Schema.from_dict` takes a dict input
explicitly, so this is needed for types to match. New tests ensure
that we behave correctly by means of UserDict.

Bump the version of mypy listed in tox config to the latest to ensure
type checking runs correctly and quickly.