Move mypy config into pyproject.toml
In #848, we introduced a `mypy.ini` config to list exceptions to mypy
rules, as that was the only way that mypy supported these listings at
the time.

Since then, mypy has added support for `tool.mypy.overrides` in
`pyproject.toml`, which allows us to consolidate configs into
`pyproject.toml`. This also provides a more compact format for listing
these exceptions in our case.