Fix the type annotation for 'argmap' (#989)
The ArgMap type alias defined the mapping option (for
`Schema.from_dict`) to allow `Field` or `type[Field]`, but this is
incorrect. The following construction does not work:

    Schema.from_dict({"a": fields.Str})

Correct by simply removing `type[Field]` from the alias.

A quick included improvement is to make the handling of type narrowing
for `argmap` a little bit more uniform in style (we had it 3 different
ways in different places), and to use a cast to replace a type ignore
with something a little easier to read.